moving functions to their rightful place
